requirements.txt
# LLM 框架
langchain>=0.1.0
langchain-core>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-chroma>=0.1.0
langchain-community>=0.0.20

# 文档处理
unstructured[all-docs]>=0.12.0
pdf2image>=1.16.3
pypdf>=3.17.0
python-docx>=1.1.0

# 向量数据库
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 工具库
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# 图像处理
Pillow>=10.2.0
pytesseract>=0.3.10

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.29.0

# RAG评估框架
ragas>=0.1.0,<1.0.0
langsmith>=0.1.0

# 可视化
matplotlib>=3.7.0

